moved js js files in root to new file
